+++ frontend/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router'; // CLI imports router
 import { ChatTemplateComponent } from './main/pages/chat-template/chat-template.component';
-import { ChatComponent } from './main/pages/chat-template/components/chat/chat.component';
+import { ChatComponent } from './main/pages/chat/chat.component';
 import { IntroductionComponent } from './main/pages/introduction/introduction.component';
 import { LoginComponent } from './main/pages/login/login.component';
