--- frontend/src/app/main/pages/chat-template/chat-template-routing.module.ts
@@ -1,13 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-const routes: Routes = [
-
-];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
-})
-
-export class ChatsTemplateRoutingModule {}
