+++ frontend/src/app/main/pages/chat/chat.component.ts
